Mapping details about credentials and success from a complex field in the table of graduation rates.

Mapping details about credentials and success from a complex field in the table of graduation rates.

Usage

GLOSSARIES_FOR_GR_TYPES

Format

GLOSSARIES_FOR_GR_TYPES

A list of three data frames.

Cohort

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 28 rows and 4 columns.

GRTYPE

<int> a code that maps to the rest of the columns

Years

<int> Whether the cohort comes from a 2- or 4-year institution.

Award Sought

<fct> Cohorts are defined by their institution and the type of award their students enrolled to earn. Values include "Certificate," "Associate's," "Bachelor's," or "Any."

Statistic

<fct> The quantity reported in rows with this GRTYPE code. Possible values are: "Everyone", "Excluded from cohort", "Adjusted cohort", "Transferred out", "Still enrolled", "No longer enrolled".

Completers

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 18 rows and 5 columns.

GRTYPE

<int> a code that maps to the rest of the columns

Years

<int> Whether the cohort comes from a 2- or 4-year institution.

Award Sought

<fct> Cohorts are defined by their institution and the type of award their students enrolled to earn. Values include "Certificate", "Associate's", "Bachelor's", or "Any degree"

Award Level

<chr> the actual credential a person earned, with the same options as above

Time to Award

<fct> Relative to the normal completion time. Values are "<=100%", "==125%", "<=150%", and "==150%".

Assistance

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 4 rows and 3 columns.

PSGRTYPE

<int> a code that maps to the rest of the columns

Years

<int> Whether the cohort comes from a 2- or 4-year institution.

Award Sought

<fct> Cohorts are defined by their institution and the type of award their students enrolled to earn. Values include "Certificate", "Associate's", "Bachelor's", or "Any degree"

Source

https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx