Definitions of columns found in data that describes postsecondary completers

Definitions of columns found in data that describes postsecondary completers

Usage

GR_VARIABLES

Format

GR_VARIABLES

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 74 rows and 8 columns.

varNumber

<int> a code for mapping from raw data to this table.

varName

<chr> a seven-character name for each variable

varTitle

<chr> a human-readable name for each variable

DataType

<chr> "N" for data and "A" for metadata

format Time

<dbl> either "Cont" (count or continuous) or "Disc" (numeric or character categories)

Race/Ethnicity

<chr> The OMB racial/ethnic category for people counted by this variable

Sex

<chr> The OMB gender category for people counted by this variable

longDescription

<chr> a verbose explanation of the variable

Source

https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx