Search a path for files with IPEDS-type names

Files that are downloaded from IPEDS typically have a constant format within each type of table. The only difference among files is a year number. For example, a file for graduation rates from 1998 would be "GR1998.xlsx"

Usage

find_yearly_tables(.path, .prefix, .suffix = "", ...)

Arguments

.path

<chr> the path to search for.

.prefix

<chr> the prefix for this type of table

.suffix

<chr?> an optional suffix to file names. Defaults to ""

...

Arguments passed on to base::dir

all.files

a logical value. If FALSE, only the names of visible files are returned (following Unix-style visibility, that is files whose name does not start with a dot). If TRUE, all file names will be returned.

full.names

a logical value. If TRUE, the directory path is prepended to the file names to give a relative file path. If FALSE, the file names (rather than paths) are returned.

recursive

logical. Should the listing recurse into directories?

ignore.case

logical. Should pattern-matching be case-insensitive?

include.dirs

logical. Should subdirectory names be included in recursive listings? (They always are in non-recursive ones).

no..

logical. Should both "." and ".." be excluded also from non-recursive listings?

Value

<chr[]> a vector of file names, with names set to matching years