Load a table of definitions for IPEDS CIP codes.

The Classification of Instructional Programs is a taxonomy of higher education programs/majors/degrees/etc. It is maintained by the National Center for Education Statistics

Usage

read_cip_codes(.path)

Arguments

.path

<chr> the full path to a file downloaded from NCES.

Value

a tibble::tibble() with five variables

Family

<chr> a two-digit code for the broadest categories of programs

Code

<chr> a 2-, 4-, or 6-digit code in ##[.####] format, family first.

Title

<chr> a short-ish label for the group or program

Definition

<chr> a longer description of the group or program

Type

<chr> whether the code has 2, 4, or 6 digits