Load IPEDS data about graduation rates

The graduation data come in two tables. One reports graduation numbers by race/ethnicity/immigration status and sex. The other reports graduation by type of financial assistance. Both tables report numbers of students and graduates, and break down by degree sought and granting institution. Each file describes one graduation year, so the enrollment year is different for 2- and 4-year programs.

Usage

read_graduates(.dem_file, .aid_file)

Arguments

.dem_file

<chr> the full path to the file with graduation and demographic information.

.aid_file

<chr> the full path to the file with graduation and financial aid information.

Value

#' An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame). See Details.

Details

Graduation rates by demographic groupings

These data include many different breakdowns, including degree sought, actual degree earned, several different times to completion and demographic information such as race/ethnicity, immigration status, and gender. When you call this function with the "demographic" table option, the output is a data frame with 15 columns.

UNITID

<int> a seven digit unique identifier for a school

Index

<int> the

Institution Level

<chr>

Credential Sought

<chr>

Credential Earned

<chr>

Relative Time

<chr>

Race/Ethnicity

<chr>

Sex

<chr>

Adjusted Cohort

<int>

Completed

<int>

Transferred

<int>

Full Cohort

<int>

Still Enrolled

<int>

No Longer Enrolled

<int>

Excluded

<int>

Graduation rates by financial assistance

These data all deal with graduation within 150% of normal completion time. That means 6 years for bachelor's programs and 3 years for associate's. When you call this function with the "financial" table option, the output is a data frame with 9 columns.

UNITID

<int> a seven digit unique identifier for a school

PSGRTYPE

<int> a one-digit code describing the type of credential earned

Assistance

<chr> the type of assistance that students receive, if any

Revised Cohort

<int> an overestimate of the number of students who could have graduated

Exclusions

<int> the number of students who should be removed from the denominator

Adjusted Cohort

<int> a better estimate of the number of students who could have graduated

Bachelor's Degrees

<int> the number of students who earned a 4-year degree

Associate's Degrees

<int> the number of students who earned a degree that took fewer than 4 years

All Degrees

<int> the total number of undegraduate degrees earned

See also