Read data about graduation and financial assistance

IPEDS tables with names like "GRXXXX_PELL_SSL" contain data about postsecondary completion and publicly-funded financial assistance.

Usage

read_graduates_and_assistance(.file)

Arguments

.file

the full path to the file

Value

An object, of class spec_tbl_df (inherits from tbl_df, tbl, data.frame), with 9 columns.

UNITID

<int> a seven digit unique identifier for a school

Years

<int> Whether the cohort comes from a 2- or 4-year institution.

Award Sought

<fct> Cohorts are defined by their institution and the type of award their students enrolled to earn. Values include "Certificate", "Associate's", "Bachelor's", or "Any degree"

Assistance

<chr> the type of assistance that students receive, if any

Cohort

<tbl> information about folks seeking credentials. See Details

Completers

<tbl> information about folks earning credentials. See Details

Details

All of these completion numbers report 150% graduation rates. There are two list-columns that contain nested data frames. Their structures are:

Cohort

Statistic

<chr> the quantity being measured, like cohort size or number of students excluded.

Count

<chr> the corresponding number of students.

Completers

Award Level

<chr> one of "Associate's Degrees", "Bachelor's Degrees", or "All Degrees".

Count

<int> the corresponding number of degrees earned

See also