Expand a compact coded variable into its multiple meanings in a lookup table

Many situations arise where a single variable, like "GRTYPE" in IPEDS, actually encodes a specific combination of values from several other variables. This function is a utility for adding all of those meanings to a table that originally just has the compact one.

Usage

unpack_variable(
  .table_with_variable,
  .variable_definitions,
  .variable_name,
  ...
)

Arguments

.table_with_variable

<tbl> the table where the variable appears without elaboration

.variable_definitions

<tbl> the table of definitions for the variable

.variable_name

<any> the name of the variable field, which should be the same in both of the other inputs

...

<chr> fields from variable_definitions to keep. Leaving this empty means "use all columns in .variable_definitions"

Value

<tbl> a data frame with all columns in .table_with_variable (except .variable_name) and the ... columns from .variable_definitions.