Read data about college completers from one zipped file

Each file contains the number of students receiving a degree or certificate. Data cover awards granted between July 1, .year - 1 and June 30, .year.

Usage

read_completers(.path, .spec)

Arguments

.path

<chr> the full path to a csv file with completer data.

.spec

<tbl> either "A," "B," or "C."

Value

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame). The exact nature of the value depends upon .spec. See Details

Details

There are three possible aggregations of completer data. The type of aggregation determines the specific layout of the return value.

Aggregation A

This aggregation reports about graduates by field of study, primary major, gender, race/ethnicity, immigration status, and type of degree. Choosing this aggregation results in a return value with 7 columns.

UNITID

<int> the seven-digit code for an institution

CIPCODE

<chr> a 2-, 4-, or 6-digit code for a field of study

MAJORNUM

<int> whether this was a first or second major

Completers

<int> the number of degrees awarded in a calendar year

Population

<chr> the race/ethnicity or other demographic category of people counted by this row

Gender

<chr> all, female, or male

Award Level

<chr> Associate's, Bachelor's, Certificate, etc.

Aggregation B

This aggregation reports about graduates by gender, race/ethnicity, and immigration status. Choosing this aggregation results in a return value with 4 columns.

UNITID

<int> the seven-digit code for an institution

Completers

<int> the number of degrees awarded in a calendar year

Population

<chr> the race/ethnicity or other demographic category of people counted by this row

Gender

<chr> all, female, or male

Aggregation C

This version of the table enumerates students by level of award and by race/ethnicity, gender and age categories. Each file contains multiple records per institution. Each record is uniquely defined by the variables IPEDS ID (UNITID) and award level (AWLEVELC ). The 'C' is stripped from "AWLEVELC" during post-read processing. Each record will contain:

Choosing this aggregation results in a return value with 7 columns.

UNITID

<int> the seven-digit code for an institution

Completers

<int> the number of degrees awarded in a calendar year

Population

<chr> the race/ethnicity or other demographic category of people counted by this row

Gender

<chr> all, female, or male

Lower Age

<int> the lowest age in the age bracket described by this row

Upper Age

<int> the highest age in the age bracket described by this row

Award Level

<chr> Associate's, Bachelor's, Certificate, etc.

Numbers are not available for combinations of race, gender and age. Beginning with 2019-20, "certificate of less than 1 year" was replaced by the following 2 categories:

certificates of less than 12 weeks

certificates of at least 12-weeks but less-than 1-year.